+++ crates/renderd-proto/src/lib.rs
@@ -1,1 +1,5 @@
 //! `Protobuf` types for Renderd control plane.
+
+pub mod generated;
+
+pub use generated::renderd::*;